old/api/requirements.txt
flask
flask_pymongo
pymongo
pyyaml
openai
flask-cors
python-docx
mongoengine
json_repair